streamlit_app/requirements.txt
streamlit==1.41.0
functions-framework==3.8.1
requests==2.32.3
pandas==2.2.2
python-dotenv==1.0.1
google-cloud-bigquery==3.20.1
